still digging in the dust

+++ src/agrum/learning/scores_and_tests/score.h
@@ -69,7 +69,7 @@ namespace gum {
      * observed countings if you are developping a new score class, or use
      * method score to get the computed score if you are an end user. */
     template <typename IdSetAlloc = std::allocator<unsigned int>,
-              typename CountAlloc = std::allocator<float>>
+              typename CountAlloc = std::allocator<double>>
     class Score : private Counter<IdSetAlloc, CountAlloc> {
       public:
       // ##########################################################################
@@ -158,7 +158,7 @@ namespace gum {
       using Counter<IdSetAlloc, CountAlloc>::setMaxNbThreads;
 
       /// returns the score corresponding to a given nodeset
-      virtual float score( unsigned int nodeset_index ) = 0;
+      virtual double score( unsigned int nodeset_index ) = 0;
 
       /// indicates whether the apriori is compatible (meaningful) with the
       /// score
@@ -201,7 +201,7 @@ namespace gum {
 
       protected:
       /// 1 / log(2)
-      const float _1log2{M_LOG2E};
+      const double _1log2{M_LOG2E};
 
       /// the a priori used by the score
       Apriori<IdSetAlloc, CountAlloc>* _apriori;
@@ -244,21 +244,21 @@ namespace gum {
        * nodes of the conditioning set (in the order in which they were
        * specified
        * when callind addNodeset, and then the target nodes. */
-      const std::vector<float, CountAlloc>&
+      const std::vector<double, CountAlloc>&
       _getAllApriori( unsigned int index );
 
       /// returns the apriori vector for a conditioning set
-      const std::vector<float, CountAlloc>&
+      const std::vector<double, CountAlloc>&
       _getConditioningApriori( unsigned int index );
 
       /// indicates whether a score belongs to the cache
       bool _isInCache( unsigned int nodeset_index ) const noexcept;
 
       /// inserts a new score into the cache
-      void _insertIntoCache( unsigned int nodeset_index, float score );
+      void _insertIntoCache( unsigned int nodeset_index, double score );
 
       /// returns a cached score
-      float _cachedScore( unsigned int nodeset_index ) const noexcept;
+      double _cachedScore( unsigned int nodeset_index ) const noexcept;
 
       /// indicates whether we use the cache or not
       bool _isUsingCache() const noexcept;
@@ -280,7 +280,7 @@ namespace gum {
       std::vector<bool> __is_cached_score;
 
       /// the vector of scores for the current nodesets
-      std::vector<float> __cached_score;
+      std::vector<double> __cached_score;
 
       /// has the a priori been computed
       bool __apriori_computed{false};
