+++ src/agrum/learning/scores_and_tests/scoreAIC.h
@@ -64,7 +64,7 @@ namespace gum {
      * details.
      */
     template <typename IdSetAlloc = std::allocator<unsigned int>,
-              typename CountAlloc = std::allocator<float>>
+              typename CountAlloc = std::allocator<double>>
     class ScoreAIC : public Score<IdSetAlloc, CountAlloc> {
       public:
       // ##########################################################################
@@ -108,7 +108,7 @@ namespace gum {
       /// @{
 
       /// returns the score corresponding to a given nodeset
-      virtual float score( unsigned int nodeset_index ) final;
+      virtual double score( unsigned int nodeset_index ) final;
 
       /// indicates whether the apriori is compatible (meaningful) with the
       /// score
@@ -153,7 +153,7 @@ namespace gum {
        * method isAprioriCompatible (the method needs be updated to take it into
        * account). */
       static bool isAprioriCompatible( const std::string& apriori_type,
-                                       float weight = 1.0f );
+                                       double weight = 1.0f );
 
       /// indicates whether the apriori is compatible (meaningful) with the
       /// score
