+++ src/agrum/learning/scores_and_tests/scoreBDeu.h
@@ -75,7 +75,7 @@ namespace gum {
      * details.
      */
     template <typename IdSetAlloc = std::allocator<unsigned int>,
-              typename CountAlloc = std::allocator<float>>
+              typename CountAlloc = std::allocator<double>>
     class ScoreBDeu : public Score<IdSetAlloc, CountAlloc> {
       public:
       // ##########################################################################
@@ -127,7 +127,7 @@ namespace gum {
       /// @{
 
       /// returns the log2(BDeu score) corresponding to a given nodeset
-      float score( unsigned int nodeset_index );
+      double score( unsigned int nodeset_index );
 
       /// indicates whether the apriori is compatible (meaningful) with the
       /// score
@@ -178,7 +178,7 @@ namespace gum {
        * method isAprioriCompatible (the method needs be updated to take it into
        * account). */
       static bool isAprioriCompatible( const std::string& apriori_type,
-                                       float weight = 1.0f );
+                                       double weight = 1.0f );
 
       /// indicates whether the apriori is compatible (meaningful) with the
       /// score
@@ -226,7 +226,7 @@ namespace gum {
       internalApriori() const noexcept final;
 
       /// sets the effective sample size of the internal apriori
-      void setEffectiveSampleSize( float ess );
+      void setEffectiveSampleSize( double ess );
 
       /// @}
 
@@ -238,7 +238,7 @@ namespace gum {
       ScoreInternalBDeuApriori<IdSetAlloc, CountAlloc> __internal_apriori;
 
       /// the effective sample size of the internal apriori
-      float __ess{1.0f};
+      double __ess{1.0f};
     };
 
   } /* namespace learning */
