+++ src/agrum/learning/scores_and_tests/scoreLog2Likelihood.tcc
@@ -81,7 +81,7 @@ namespace gum {
     /// indicates whether the apriori is compatible (meaningful) with the score
     template <typename IdSetAlloc, typename CountAlloc>
     bool ScoreLog2Likelihood<IdSetAlloc, CountAlloc>::isAprioriCompatible(
-        const std::string& apriori_type, float weight ) {
+        const std::string& apriori_type, double weight ) {
       // check that the apriori is compatible with the score
       if ( ( apriori_type == AprioriDirichletType::type ) ||
            ( apriori_type == AprioriSmoothingType::type ) ||
@@ -121,7 +121,7 @@ namespace gum {
 
     /// returns the score corresponding to a given nodeset
     template <typename IdSetAlloc, typename CountAlloc>
-    float ScoreLog2Likelihood<IdSetAlloc, CountAlloc>::score(
+    double ScoreLog2Likelihood<IdSetAlloc, CountAlloc>::score(
         unsigned int nodeset_index ) {
       // if the score has already been computed, get its value
       if ( this->_isInCache( nodeset_index ) ) {
@@ -129,10 +129,10 @@ namespace gum {
       }
 
       // get the counts for all the targets and for the conditioning nodes
-      const std::vector<float, CountAlloc>& N_ijk =
+      const std::vector<double, CountAlloc>& N_ijk =
           this->_getAllCounts( nodeset_index );
       const unsigned int targets_modal = N_ijk.size();
-      float score = 0;
+      double score = 0;
 
       // get the nodes involved in the score
       const std::vector<unsigned int, IdSetAlloc>* conditioning_nodes =
@@ -142,14 +142,14 @@ namespace gum {
       // without conditioning nodes
       if ( conditioning_nodes ) {
         // get the counts for all the targets and for the conditioning nodes
-        const std::vector<float, CountAlloc>& N_ij =
+        const std::vector<double, CountAlloc>& N_ij =
             this->_getConditioningCounts( nodeset_index );
         const unsigned int conditioning_modal = N_ij.size();
 
         if ( this->_apriori->weight() ) {
-          const std::vector<float, CountAlloc>& N_prime_ijk =
+          const std::vector<double, CountAlloc>& N_prime_ijk =
               this->_getAllApriori( nodeset_index );
-          const std::vector<float, CountAlloc>& N_prime_ij =
+          const std::vector<double, CountAlloc>& N_prime_ij =
               this->_getConditioningApriori( nodeset_index );
 
           // compute the score: it remains to compute the log likelihood, i.e.,
@@ -157,15 +157,15 @@ namespace gum {
           // equivalent to:
           // sum_j=1^q_i sum_k=1^r_i N_ijk log N_ijk - sum_j=1^q_i N_ij log N_ij
           for ( unsigned int k = 0; k < targets_modal; ++k ) {
-            const float new_count = N_ijk[k] + N_prime_ijk[k];
+            const double new_count = N_ijk[k] + N_prime_ijk[k];
             if ( new_count ) {
-              score += new_count * logf( new_count );
+              score += new_count * std::log( new_count );
             }
           }
           for ( unsigned int j = 0; j < conditioning_modal; ++j ) {
-            const float new_count = N_ij[j] + N_prime_ij[j];
+            const double new_count = N_ij[j] + N_prime_ij[j];
             if ( new_count ) {
-              score -= new_count * logf( new_count );
+              score -= new_count * std::log( new_count );
             }
           }
         } else {
@@ -175,12 +175,12 @@ namespace gum {
           // sum_j=1^q_i sum_k=1^r_i N_ijk log N_ijk - sum_j=1^q_i N_ij log N_ij
           for ( unsigned int k = 0; k < targets_modal; ++k ) {
             if ( N_ijk[k] ) {
-              score += N_ijk[k] * logf( N_ijk[k] );
+              score += N_ijk[k] * std::log( N_ijk[k] );
             }
           }
           for ( unsigned int j = 0; j < conditioning_modal; ++j ) {
             if ( N_ij[j] ) {
-              score -= N_ij[j] * logf( N_ij[j] );
+              score -= N_ij[j] * std::log( N_ij[j] );
             }
           }
         }
@@ -198,40 +198,40 @@ namespace gum {
         // here, there are no conditioning nodes
 
         // get the counts for all the targets and for the conditioning nodes
-        const std::vector<float, CountAlloc>& N_ijk =
+        const std::vector<double, CountAlloc>& N_ijk =
             this->_getAllCounts( nodeset_index );
         const unsigned int targets_modal = N_ijk.size();
 
         if ( this->_apriori->weight() ) {
-          const std::vector<float, CountAlloc>& N_prime_ijk =
+          const std::vector<double, CountAlloc>& N_prime_ijk =
               this->_getAllApriori( nodeset_index );
 
           // compute the score: it remains to compute the log likelihood, i.e.,
           // sum_k=1^r_i N_ijk log (N_ijk / N), which is also
           // equivalent to:
           // sum_j=1^q_i sum_k=1^r_i N_ijk log N_ijk - N log N
-          float N_plus_N_prime = 0;
+          double N_plus_N_prime = 0;
           for ( unsigned int k = 0; k < targets_modal; ++k ) {
-            const float new_count = N_ijk[k] + N_prime_ijk[k];
+            const double new_count = N_ijk[k] + N_prime_ijk[k];
             if ( new_count ) {
-              score += new_count * logf( new_count );
+              score += new_count * std::log( new_count );
               N_plus_N_prime += new_count;
             }
           }
-          score -= N_plus_N_prime * logf( N_plus_N_prime );
+          score -= N_plus_N_prime * std::log( N_plus_N_prime );
         } else {
           // compute the score: it remains to compute the log likelihood, i.e.,
           // sum_k=1^r_i N_ijk log (N_ijk / N), which is also
           // equivalent to:
           // sum_j=1^q_i sum_k=1^r_i N_ijk log N_ijk - N log N
-          float N = 0;
+          double N = 0;
           for ( unsigned int k = 0; k < targets_modal; ++k ) {
             if ( N_ijk[k] ) {
-              score += N_ijk[k] * logf( N_ijk[k] );
+              score += N_ijk[k] * std::log( N_ijk[k] );
               N += N_ijk[k];
             }
           }
-          score -= N * logf( N );
+          score -= N * std::log( N );
         }
 
         // divide by log(2), since the log likelihood uses log_2
